Minor changes to TextBlock
Add a new parameter "symbolic_names" to TextBlock._as_ascii which defaults
to false. When false, character codes that don't map to a single character,
or which aren't escape sequences (i.e.: \n), output an escape sequence for it.

For example: T -> T
But: "(Star)" -> \u819A

This is so strings can be read, modified to replace a substring, and then
submitted again to update the string, which is important for handling the
shuffling of NPCs while keeping the text mapped to the location they are.