Update TextBlock a bit.
Based on the newer way I wrote the ShopData out, updated TextBlock
to use the same method. That is, rather than return a tuple for
the caller to patch itself, create a routine that takes a rom,
and TextBlock creates its own patch and applies both the LUT
and block changes at once, returning the new rom.