Fix the caravan shop.
There's something that breaks when writing all the maps end to end.
For now, this just writes maps back over themselves, which works
at the moment, since we're not actually changing them, but it would
be nice to better understand what else is in the map memory area
that's getting overwritten.

Or if there's just something special about map 0x73 ¯\_(ツ)_/¯