Fix Citadel of Trials entrance.
Fixes #9

In the NES version, the game says that you must possess the Crown
to enter Castle Ordeal. In this version, a different flag is used
that's set sometime after the Crown chain, but before arriving at
the entrance.

This changes it so that as soon as you receive the Crown (and exit
to the world map), that flag is set.