Get e2s.py in better shape.
I *think* e2s.py should be in good enough shape now to work with most of
the map and core game events.

It does not provide as much information as edump.py, but it creates text
that will be able to be parsed and assembled into a new event by the
assembler.