First events rewritten
This commits the event preprocessor that supports #define
and #ifdef statements (though nothing takes advantage of #ifdef
at the moment.)

The event preprocessor and event assembler work together to rewrite
the King and Sara's events as a proof-of-concept. The "King"
is currently forced to give the Lute and with the seed I'm using
(fadebabe) Sara was replace by Dr. Unne.

The other events will need to be translated to this system.