Toss chest events.
Chest events are now rewritten to be NPC events.

Known issue: The chests aren't simply "chests" in the map, they're
also drawn on the floor, so placing a non-chest NPC over it looks
odd. This will be addressed by rewriting parts of the map so
the chest "background" isn't there.