Dump non-command tokens.
There are a bunch of special case values for various commands, such as
closing a dialog, but having special tokens for them just clutters up
the grammar.

This switches those values to regular values, which means that it's
possible to use other values for those commands now (if it makes sense),
or to just use the #define values from the standard header.