Finally add IPS patching
This changes the way the randomizer delivers the ROM. Rather than
uploading your ROM, which is randomized and returned (16mb of data
up and back, most of which is the same), the settings are passed
to the server, which creates a patch for the ROM. This patch is
applied to your ROM locally and you're given the option to save it.

The old method still exists for time being, and is available via
the [legacy site](hmsjayne.com/legacy).