"Better" shop inventory
This changes it so weapon and armor sold in shops is unique (that is, if
a Mythril Knife is for sale in Elfheim it will only be sold there).

Item shops do _not_ only sell unique items, since it makes more sense to
have multiple shops selling Potions, for example.

This also ensures Antidoes and Golden Needles are for sale in Cornelia,
and makes a change to make it more likely that there are more weapons
and pieces of armor for sale in general.