Start with the Bridge too
This updates the code to give the bridge at the start of the game, since
it's otherwise completely useless. Like Bikke, the King now gives a
random piece of gear, which may or may not be useful.

A side effect of the way this change was implemented makes disabling key
item shuffle work like one might expect: no ship or bridge is given at
the start, and so progression is completely vanilla.