Updates
A few updates for things. First, this tests a change to use the more vanilla
approach to event handling. It works better which is nice, because that means
I can get the dialog to work better.

It's a bit annoying because I need to export and re-update all the scripts to
use this, and this could introduce bugs again that took awhile to work out
originally.

On the other hand, this will let me change some of the script commands to be
a bit easier to reason about, like replacing "jz" to something like
"if_not_set_goto" or whatever like that.

(Or even: "goto <label> if %flag set")