Progress
Lots more commands are documented and disassembled now.
But they can't be reassembled, so... there's that.

On the other hand, a lot of them will be removed to make the events
more speedy for the randomizer, so.... It's mostly putting things
in so that I can more easily see what's happening and cut out stuff.