Start works!
Moved "FREE_START" into the Cornelia Guards event and made
it very short (without teleporting).

Also added it to Chaos Shrine, so that the door is unlocked
and the FREE_START is given there, if the player decides
to skip Cornelia and run directly there.

Updated Chaos Shrine to also check for the Earth Crystal
since, in theory, it may be possible to light the others
without that one, and we don't want to let people skip
it. =)