Event encounters working :)
"Stole" 3 encounters from the DoS content to use for the encounters
to get the items in Marsh Cave, Ice Cavern, and the Citadel of Trials.

0x101 (Cerberus) -> 0x1c (Piscodemons)
0x102 (Ahriman) -> 0x69 (Evil Eye) ... I mean, obviously
0x103 (2-Headed Dragon) -> 0x4b (Zombie Dragon) ... again :)

This means that the original encounters _could_ be updated to be
runnable, or have different surprise chance. I'm not sure if
that Evil Eye encounter or Zombie Dragon appears elsewhere,
and the Piscodemons version is used elsewhere in Marsh, so...

If, in the future, it becomes possible to update the code so that
we could just send in the specific encounter we want, then the
code to revert this is (at least) pretty simple to find/remove :)